+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file is used to track changes made to the project over time.
 
+## [Unreleased]
+### Added
+- Binding for `SolveResult::cost`
+- Binding for `SolveResult::get_B`
+
 ## [0.2.0] - 2025-12-27
 ### Added
 - Python bindings for the following methods and attributes: